FIX: multi-leg fractional-qty truncation, dead FCF field, growth-trend reason codes, loader-health false positives
- reconciliation.py: reconcile_exit_fills() int()-truncated entry_qty before
  computing multi-leg P&L, understating cost basis/risk and overstating
  pnl_pct/exit_r_multiple (same bug class as executor_exit_handler.py's fix).
- sec_base.py: free_cash_flow was NULL for every cashflow-statement row (no
  direct XBRL tag exists for it) - now derived as operating_cash_flow - capex.
- load_value_quality_growth_metrics.py: growth-trend fields (net_income_growth_yoy,
  margins, roe_trend, fcf/ocf growth, etc.) now carry an unavailable_reason instead
  of a bare NULL when the prior fiscal year's row lacks the input; scores.py/
  frontend surface the reason instead of an unexplained "No data" badge.
- monitoring.py: /api/algo/freshness/extended flagged healthy loaders (status
  HEALTHY/DEPRECATED from pipeline_health.py's freshness sweep) as "unhealthy"
  under a naive status != COMPLETED check, and could silently truncate the real
  unhealthy count via LIMIT 30 before it was computed.

+++ algo/infrastructure/reconciliation.py
@@ -1620,7 +1620,13 @@ class DailyReconciliation:
                     try:
                         entry_price = float(entry_price)
                         stop_loss_price = float(stop_loss_price)
-                        entry_qty = int(entry_qty)
+                        # float (not int): algo_trades.entry_quantity is NUMERIC(18,4) - real
+                        # fractional-share entries exist in this DB. int() truncation here fed
+                        # a too-small cost basis/risk denominator into original_cost_basis/
+                        # original_risk_dollars below for multi-leg fills - same bug class
+                        # already fixed in executor_exit_handler.py's _compute_cumulative_pnl
+                        # (the synchronous exit path this function mirrors for async fills).
+                        entry_qty = float(entry_qty)
                     except (ValueError, TypeError) as e:
                         cur.execute("RELEASE SAVEPOINT reconcile_fill")
                         raise ValueError(
